% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_mismatch.R
\name{comm_mismatch}
\alias{comm_mismatch}
\title{Community ToM Metrics}
\usage{
comm_mismatch(
  data,
  alt = "two.sided",
  method = "empirical",
  emt = 5,
  base_y = 5,
  quants = c(0.025, 0.975),
  alpha = 0.05,
  ks_t = 0.6,
  nboot = 100,
  unemergence = F
)
}
\arguments{
\item{data}{A data frame containing the time series to test for ToM. Must contain year (column named 'year'), timing of phenological event of interest, in Julian day (column named 'event'), and unique identifier for each individual time series for which ToPE and ToEE should be calculated (column named 'species').}

\item{alt}{Alternative hypothesis for emergence testing. Set to 'two.sided' by default, indicating checking for sp1 event timing to to be either greater than or less than sp2.}

\item{method}{Methodology used for emergence calculations. Set to 'empirical' for empirical testing (default), or 'statistical' for statistical testing using the Kolmogorov-Smirnov test.}

\item{emt}{Number of consecutive years of positive test results required to define emergence.}

\item{base_y}{Number of years at the start of the time series to compare to ensure testing is appropraite. Species pairs will only be tested if the first base_y years event values are within 1 standard deviation.}

\item{quants}{Quantiles used for empirical testing. Unused if method = 'statistical'.}

\item{alpha}{Alpha value used to determine significance for statistical testing. Unused if method = 'empirical'.}

\item{ks_t}{Proportion of significant KS Tests required to define a positive test result. Unused if method = 'empirical'.}

\item{nboot}{Number of boostrapped KS tests to run. Unused if method = 'empirical'.}

\item{unemergence}{If F, all years after first emergence are set to emerged. If T, calculation for each individual year is returned.}
}
\value{
A data frame of species pairs with test result and emergence status for each individual year. Designed to feed into comm_mm_by_spp() and comm_mm_by_year() functions.
}
\description{
Calculate ToM metrics for all relevant species pairs in a dataset
}
\examples{
# Set seed
set.seed(123)

# Create test dataset 1
year = seq(1,30,1)
event = round(rnorm(30, 100, 5))- seq(1,30,1)
env = (rnorm(30, 15, 0.5))
species = 1
sp1 = data.frame(year, species, event)

# Create test dataset 2
year = seq(1,30,1)
event = round(rnorm(30, 100, 5))
species = 2
sp2 = data.frame(year, species, event)

# Create test dataset 3
year = seq(1,30,1)
event = round(rnorm(30, 95, 5))- seq(1,15,1)
species = 3
sp3 = data.frame(year, species, event)

# combine
dataset = as.data.frame(rbind(sp1, sp2, sp3))

# Calculate empirical time of phenological emergence (ToPE)
comm_mismatch(dataset)
}
