% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comm_mm_by_year.R
\name{comm_mm_by_year}
\alias{comm_mm_by_year}
\title{Community Mismatch by Year}
\usage{
comm_mm_by_year(output, plot = T)
}
\arguments{
\item{output}{Output from comm_mismatch().}

\item{plot}{Output data frame from community().}
}
\value{
A sorted and summarized data frame generated from the output of comm_musmatch().
}
\description{
Summarizes and plots mismatch proportion from comm_mismatch() by year.
}
\examples{
# Set seed
set.seed(123)

# Create test dataset 1
year = seq(1,30,1)
event = round(rnorm(30, 100, 5))- seq(1,30,1)
env = (rnorm(30, 15, 0.5))
species = 1
sp1 = data.frame(year, species, event)

# Create test dataset 2
year = seq(1,30,1)
event = round(rnorm(30, 100, 5))
species = 2
sp2 = data.frame(year, species, event)

# Create test dataset 3
year = seq(1,30,1)
event = round(rnorm(30, 95, 5))- seq(1,15,1)
species = 3
sp3 = data.frame(year, species, event)

# combine
dataset = as.data.frame(rbind(sp1, sp2, sp3))

# Calculate empirical time of phenological emergence (ToPE)
mm_output = comm_mismatch(dataset)

# Calculate and plot mismatch by species and year
comm_mm_by_spp(mm_output)
}
