% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_tom.R
\name{emp_tom}
\alias{emp_tom}
\title{Calculate Time of Mismatch (ToM) between two time series}
\usage{
emp_tom(
  sp1,
  sp2,
  emt = 5,
  plot = T,
  max_y = 0,
  alt = "two.sided",
  quants = c(0.025, 0.975),
  unemergence = F
)
}
\arguments{
\item{sp1}{A data frame containing the time series to test for ToM. Must contain year (column named 'year'), timing of phenological event of interest, in Julian day (column named 'event').}

\item{sp2}{A data frame containing the time series against which sp1 is compared for ToM. Must contain year (column named 'year'), timing of phenological event of interest, in Julian day (column named 'event').}

\item{emt}{Number of consecutive years of positive test results required to define emergence.}

\item{plot}{If TRUE, will generate a plot of test result against year.}

\item{max_y}{Moving year window used to generate counterfactual. If 0, moving window is deactivated, else length of moving year window.}

\item{alt}{Alternative hypothesis for emergence testing. Set to 'two.sided' by default, indicating checking for sp1 event timing to to be either greater than or less than sp2.}

\item{quants}{Quantiles of the counterfactual data used to determine the emergence threshold}

\item{unemergence}{If F, all years after first emergence are set to emerged. If T, calculation for each individual year is returned.}
}
\value{
A data frame containing year, test result (p, binary. 1 = threshold exceeded), and emergence status (binary, 1 = emerged)
}
\description{
Calculate Time of Mismatch (ToM) between two time series
}
\examples{
# Set seed
set.seed(123)

# Create test dataset
year = seq(1,30,1)
event = round(rnorm(30, 100, 5))- seq(1,30,1)
species = 1
sp1 = data.frame(year, species, event)

# Create test dataset
year = seq(1,30,1)
event = round(rnorm(30, 100, 5))
species = 2
sp2 = data.frame(year, species, event)

# Calculate empirical time of phenological emergence (ToPE)
emp_tom(sp1, sp2)
}
