% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_tope.R
\name{emp_tope}
\alias{emp_tope}
\title{Empirical Time of Phenological Emergence (ToPE)}
\usage{
emp_tope(
  data,
  emt = 5,
  plot = T,
  max_y = 0,
  alt = "greater",
  quants = c(0.25, 0.75),
  unemergence = F,
  model = "linear",
  baseline = "static",
  ...
)
}
\arguments{
\item{data}{A data frame containing the time series to test for ToEE. Must contain year (column named 'year') and timing of phenological event of interest, in Julian day (column named 'event').}

\item{emt}{Number of consecutive years of positive test results required to define emergence.}

\item{plot}{If TRUE, will generate a plot of test result against year.}

\item{max_y}{Length of baseline window used to generate detrended counterfactual. If 0 uses all years, else uses maximum of max_y years.}

\item{alt}{Alternative hypothesis for emergence testing ("two.sided", "less", or "greater"). Set to 'greater' by default, indicating checking for an decreasing trend in the time series of event ~ year.}

\item{quants}{Quantiles of the detrended data used to determine the emergence threshold}

\item{unemergence}{If F, all years after first emergence are set to emerged. If T, calculation for each individual year is returned.}

\item{model}{Type of model used for detrending. 'linear' or 'curvilinear'. 'linear' uses a simple linear model, 'curvilinear' adds an x^2 term.}

\item{baseline}{Static or moving baseline used to generate detrended counterfactual. If 'static', same baseline is used for all years tested. If 'moving', baseline moves with test year.}

\item{...}{Additional arguments to feed to \code{lm()}}
}
\value{
A data frame containing year, test result (p, binary. 1 = threshold exceeded), and emergence status (binary, 1 = emerged)
}
\description{
Calculate Time of Phenological Emergence (ToPE) for a single phenology time series using empirical methodology.
}
\examples{
# Set seed
set.seed(123)

# Create test dataset
year = seq(1,30,1)
event = round(rnorm(30, 100, 5))- seq(1,30,1)
dataset = data.frame(year, event)

# Calculate empirical time of phenological emergence (ToPE)
emp_tope(dataset)
}
