% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks.R
\name{ks}
\alias{ks}
\title{Internal function for running bootstrapped KS tests}
\usage{
ks(data1, data2, alt, nboot = 100, alpha = 0.05)
}
\arguments{
\item{data1}{test dataset}

\item{data2}{control dataset}

\item{alt}{alternative hypothesis for KS test}

\item{nboot}{number of bootstrap samples}

\item{alpha}{significance value}
}
\value{
internal output for statistical emergence tests
}
\description{
Internal function for running bootstrapped KS tests
}
\examples{
1+1
}
