% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_toee.R
\name{ks_toee}
\alias{ks_toee}
\title{Statistical Time of Environmental Emergence (ToPE)}
\usage{
ks_toee(
  data,
  emt = 5,
  alpha = 0.05,
  ks_t = 0.6,
  nboot = 100,
  plot = T,
  max_y = 0,
  alt = "less",
  unemergence = F
)
}
\arguments{
\item{data}{A data frame containing the time series to test for ToEE. Must contain year (column named 'year') and environmental condition of interest (column named 'env').}

\item{emt}{Number of consecutive years of positive test results required to define emergence.}

\item{alpha}{Significance value for KS test.}

\item{ks_t}{Proportion of significant bootstrap KS tests required for a positive test result}

\item{nboot}{Number of KS test bootstraps samples.}

\item{plot}{If TRUE, will generate a plot of test result against year.}

\item{max_y}{Moving year window used to generate detrended counterfactual. If 0, moving window is deactivated, else length of moving year window.}

\item{alt}{Alternative hypothesis for emergence testing. Set to 'greater' by default, indicating checking for an decreasing trend in the time series of event ~ year.}

\item{unemergence}{If F, all years after first emergence are set to emerged. If T, calculation for each individual year is returned.}
}
\value{
A data frame containing year, test result (p, binary. 1 = threshold exceeded), and emergence status (binary, 1 = emerged)
}
\description{
Calculate Time of Environmental Emergence (ToEE) for a single phenology time series using statistical methodology.
}
\examples{
# Set seed
set.seed(123)

# Create test dataset
year = rep(seq(1,30,1), 10)
env = (rnorm(300, 15, 0.5)) + rep(seq(0.1,3,0.1),10)
dataset = data.frame(year, env); dataset = dataset[order(dataset$year),]

# Calculate statistical time of emergence
ks_toee(dataset)
}
