% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ks_tom.R
\name{ks_tom}
\alias{ks_tom}
\title{Statistical Time of Mismatch (ToM)}
\usage{
ks_tom(
  sp1,
  sp2,
  alt = "two.sided",
  emt = 5,
  alpha = 0.05,
  ks_t = 0.6,
  nboot = 100,
  plot = T,
  unemergence = F
)
}
\arguments{
\item{sp1}{A data frame containing the time series to test for ToM. Must contain year (column named 'year'), timing of phenological event of interest, in Julian day (column named 'event').}

\item{sp2}{A data frame containing the time series against which sp1 is compared for ToM. Must contain year (column named 'year'), timing of phenological event of interest, in Julian day (column named 'event').}

\item{alt}{Alternative hypothesis for emergence testing. Set to 'two.sided' by default, indicating checking for sp1 event timing to to be either greater than or less than sp2.}

\item{emt}{Number of consecutive years of positive test results required to define emergence.}

\item{alpha}{Significance value for KS test.}

\item{ks_t}{Proportion of significant bootstrap KS tests required for a positive test result}

\item{nboot}{Number of KS test bootstraps samples.}

\item{plot}{If TRUE, will generate a plot of test result against year.}

\item{unemergence}{If F, all years after first emergence are set to emerged. If T, calculation for each individual year is returned.}
}
\value{
A data frame containing year, test result (p, binary. 1 = threshold exceeded), and emergence status (binary, 1 = emerged)
}
\description{
Calculate Time of Mismatch (ToM) between two time series use statistical methodology.
}
\examples{
# Set seed
set.seed(123)

# Create test dataset 1
year = rep(seq(1,30,1), 10)
event = round(rnorm(300, 100, 5))- rep(seq(1,30,1), 10)
species = 1
sp1 = data.frame(year, species, event); sp1 = sp1[order(sp1$year),]

# Create test dataset 2
year = rep(seq(1,30,1), 10)
event = round(rnorm(300, 100, 5))
species = 2
sp2 = data.frame(year, species, event); sp2 = sp2[order(sp2$year),]

# Calculate empirical time of phenological emergence (ToPE)
ks_tom(sp1, sp2)
}
