% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_ts.R
\name{sim_ts}
\alias{sim_ts}
\title{Phenology Time Series Simulation}
\usage{
sim_ts(data, simyears)
}
\arguments{
\item{data}{A data frame containing a phenological time series for a single species, which will be simulated forward. Must contain year (column named 'year'), timing of phenological event of interest, in Julian day (column named 'event'), environmental condition of interest (column named 'env'), and unique identifier for each individual time series for which ToPE and ToEE should be calculated (column named 'species').}

\item{simyears}{Future years to simulate.}
}
\value{
A data frame containing the original data, as well as future simulated years.
}
\description{
Simulates future phenology timing and temperature time series for all species in a dataset, assuming a constant continued linear trend. Spread is simulated from the detrended original time series. It is recommended to average the result of many iterations of this function to account for random sampling.
}
\examples{
# Set seed
set.seed(123)

# Create test dataset
year = rep(seq(1,30,1), 10)
event = round(rnorm(300, 100, 5))- rep(seq(1,30,1), 10)
species = 1
env = (rnorm(300, 15, 0.5))
dataset = data.frame(year, species, event, env); dataset = dataset[order(dataset$year),]

# Years to Simulate (10 more years)
years_sim = max(year+1):max(year+10)

# Run simulation
sim_ts(dataset, years_sim)
}
