% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/community.R
\name{community}
\alias{community}
\title{Community ToPE and ToEE Metrics}
\usage{
community(
  data,
  em_alt = "greater",
  dc_alt = "less",
  method = "empirical",
  emt = 5,
  quants = c(0.025, 0.975),
  alpha = 0.05,
  ks_t = 0.6,
  nboot = 100,
  max_y = 0,
  unemergence = F
)
}
\arguments{
\item{data}{A data frame containing the time series to test for ToEE and ToPE. Must contain year (column named 'year'), timing of phenological event of interest, in Julian day (column named 'event'), environmental condition of interest (column named 'env'), and unique identifier for each individual time series for which ToPE and ToEE should be calculated (column named 'species').}

\item{em_alt}{Alternative hypothesis for ToPE testing. Set to 'greater' by default, indicating checking for an decreasing trend in the time series of event ~ year.}

\item{dc_alt}{Alternative hypothesis for ToEE emergence testing. Set to 'less' by default, indicating checking for an increasing trend in the time series of env ~ year.}

\item{method}{Methodology used for emergence calculations. Set to 'empirical' for empirical testing (default), or 'statistical' for statistical testing using the Kolmogorov-Smirnov test.}

\item{emt}{Number of consecutive years of positive test results required to define emergence.}

\item{quants}{Quantiles used for empirical testing. Unused if method = 'statistical'.}

\item{alpha}{Alpha value used to determine significance for statistical testing. Unused if method = 'empirical'.}

\item{ks_t}{Proportion of significant KS Tests required to define a positive test result. Unused if method = 'empirical'.}

\item{nboot}{Number of boostrapped KS tests to run. Unused if method = 'empirical'.}

\item{max_y}{Moving year window used to generate detrended counterfactual. If 0, moving window is deactivated, else length of moving year window.}

\item{unemergence}{If F, all years after first emergence are set to emerged. If T, calculation for each individual year is returned.}
}
\value{
A data frame containing classification results for each species and year determined by ToPE and ToEE test results. Designed to feed into class_by_year() and class_by_species() functions.
}
\description{
Calculate ToPE and ToEE metrics for all species in a dataset
}
\examples{
# Set seed
set.seed(123)

# Create test dataset 1
year = seq(1,30,1)
event = round(rnorm(30, 100, 5))- seq(1,30,1)
env = (rnorm(30, 15, 0.5))
species = 1
sp1 = data.frame(year, species, event, env)

# Create test dataset 2
year = seq(1,30,1)
event = round(rnorm(30, 100, 5))
env = (rnorm(30, 15, 0.5)) + seq(0.1,3,0.1)
species = 2
sp2 = data.frame(year, species, event, env)

# combine
dataset = as.data.frame(rbind(sp1, sp2))

# Calculate empirical time of phenological emergence (ToPE)
community(dataset)
}
