% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp_tope.R
\name{emp_tope}
\alias{emp_tope}
\title{Calculate Time of Phenological Emergence (ToPE) for a single phenology time series}
\usage{
emp_tope(
  data,
  emt = 5,
  plot = T,
  max_y = 0,
  alt = "greater",
  quants = c(0.025, 0.975),
  unemergence = F
)
}
\arguments{
\item{data}{A data frame containing the time series to test for ToEE. Must contain year (column named 'year') and timing of phenological event of interest, in Julian day (column named 'event').}

\item{emt}{Number of consecutive years of positive test results required to define emergence.}

\item{plot}{If TRUE, will generate a plot of test result against year.}

\item{max_y}{Moving year window used to generate detrended counterfactual. If 0, moving window is deactivated, else length of moving year window.}

\item{alt}{Alternative hypothesis for emergence testing. Set to 'greater' by default, indicating checking for an decreasing trend in the time series of event ~ year.}

\item{quants}{Quantiles of the detrended data used to determine the emergence threshold}

\item{unemergence}{If F, all years after first emergence are set to emerged. If T, calculation for each individual year is returned.}
}
\value{
A data frame containing year, test result (p, binary. 1 = threshold exceeded), and emergence status (binary, 1 = emerged)
}
\description{
Calculate Time of Phenological Emergence (ToPE) for a single phenology time series
}
\examples{
# Set seed
set.seed(123)

# Create test dataset
year = seq(1,30,1)
event = round(rnorm(30, 100, 5))- seq(1,30,1)
dataset = data.frame(year, event)

# Calculate empirical time of phenological emergence (ToPE)
emp_tope(dataset)
}
