% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenometrics-package.R
\docType{package}
\name{phenometrics-package}
\alias{phenometrics}
\alias{phenometrics-package}
\title{phenometrics: Time of Emergence Metrics and Community-Level Indicators for Phenological Responses to Climate Change}
\description{
This package contains functions to analyze phenological responses to climate change using Time of Emergence (ToE) methodology, which is designed to identify the point at which a species establishes a 'new normal' in terms of its phenology. It also contains functions to summarize the ToE results of many species into community-level indicators, and test for match/mismatch dynamics between species.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
